\name{ChangeAnomalyDetection-package}
\alias{ChangeAnomalyDetection-package}
\alias{ChangeAnomalyDetection}
\docType{package}
\title{
ChangeAnomalyDetection Package
}
\description{
ChangeAnomalyDetection Package
}
\details{
\tabular{ll}{
Package: \tab ChangeAnomalyDetection\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2012-09-29\cr
License: \tab What license is it under?\cr
}
}
\author{
Maintainer: Yohei Sato <yokkuns@tkul.jp>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
